requirements.txt
alembic==1.13.2
asgiref==3.8.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.1
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
platformdirs==4.2.2
PyMySQL==1.1.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.4
